Various changes to the cabcalc GUI & refactoring of gui.py.
Changed `Customer' entry widget to `Description' and made it optional.
Added output buttons: `Save Cutlist' and `Optimize Panel Layout'.
Added a `job' attribute to the Application class to hold the job object with
all the information about the current job, to make it more persistent. This
became necessary, as it is created in the calculate_job() function but not used
until the save_cutlist() function is called when the `Save Cutlist' button is
clicked.