Sundry changes to the cutlist-platy.py module.
Added Lee's isometric view drawing code, modified so that it produces a Drawing
that can be added to the PLATYPUS document, as the cutlist doc is now a PLATYPUS
doc. Added tables to the header frame and righthand column for the panel
drawings. Added more detail to the panel drawings. Made the Parts List use fixed
width font, and changed from using Paragraphs to using XPreformatted flowables,
so that the whitespace is not removed.