In cabcalc.py, moved file-local var specifying utf-8 to second line.
Putting `-*- coding: utf-8 -*-' after the `#!/usr/bin/env python3' on
the first line prevents the program from being run like a regular
binary because it thinks the `-*- coding:...' is part of the file name
to execute.

However, luckily, if the first line contains a shebang (#!...)
construction, Emacs knows that file-local variables cannot be specified
on the first line and so it will look for them on the second line.