Specify fillers via an enumeration (neither left right both) instead of an int.
To specify fillers, we now choose which ends have fillers (neither, left, right
or both), rather than specifying the number of fillers, 0, 1 or 2. The command
line option -f --fillers now takes `neither', `left', `right' or `both' and
converts it to <Ends.neither: 1>, <Ends.left: 2> , <Ends.right: 3>, etc.

The GUI now has radio buttons for `neither', `left', `right' or `both' instead
of for 0, 1 or 2.

All of this allows us to fix the bug from issue #11, by properly displaying at
the right end of the cutlist header whether the left and right ends are finished
or unfinished.