Use separate primary and door material/thickness in the GUI.
Added widgets to accommodate the two different sets of material/thickness, as
well as code to hook them up and make thickness change when material is changed.

Modified as well the calculate_job function to handle the different primary/door
material and thickness when creating the Run object to calculate the run.