Fix spacing of isometric cabinet drawing on cutlist.
Added code to calculate properly the width and height required for the drawing,
then used this to specify the size when creating the Drawing object, so that when
it is flowed into the document it won't overlap anything else.

Set the hAlign attribute of the isometric Drawing object, which is a flowable,
to 'CENTER' so the drawing is centered in its column (frame).

Added a Spacer before the isometric drawing in the elements list.

Changed variable names so it is clearer what their function is and how they are
derived. Added comments to document changes to the code.

Changed the computation of isoNlr45 to use job.cabs.topnailer_depth, in order to
be more explicit for readability.

Replaced the variable `nlr' with `isoNlr45' wherever it was used, as their
values should have been the same, since they are meant to be the horizontal and
vertical projections, respectively, of the angled topnailer depth lines. Since
these are angled at 45 degrees, the horizontal and vertical projections must be
equal. However, in actuality they were not. `nlr' was just a bit larger in
value, 2 versus 2 * sin(45). This is what caused the horizontal line for the
front top edge of the front topnailer to be misaligned, since it was placed too
high. This change fixes that problem; the lines are now perfectly aligned, even
when blown up to 400%.