Remove defaults from boundsln_len formal arguments; pass in explicit values.
Passing in explicit values for the dimension bounds line lengths enhances
clarity. Some drawings use a bounds line length of 10 and some use 14. This way
it is easy to see which we are using when. Before when sometimes not passing in
a value for the bounds line length and allowing it to default, it was not even
clear that we were using two different values for the length.