Fixed isoLines so that back panel appears proper thickness, not too thick.
This just required multiplying the back panel thickness, which was subtracted,
by sin(45).

Also changed some quantities from cabs.prim_thickness to cabs.back_thickness or
cabs.side_thickness, as appropriate. This did not make any difference in the
values computed right now, as side and back thickness are the same as the
primary thickness, but for future proofing against any changes in the code, this
is the right thing to do.