Add option for cabinet legs and different bottom thickness.
The dictionary of material thicknesses is changed from just including one
thickness per material to including a pair of thicknesses for each material,
the first thickness is the usual one, and the second is for bottom panels when
legs will be attached, which need to be thicker than 3/4" so that the leg screws
can grab.

The Run class has a new bottom_thickness attribute, which can be different from
the primary material thickness, and a new has_legs property attribute, which
when changed fires a function which will change the bottom thickness if necessary.

If the bottom thickness is different from the primary material thickness, the
job specification material information section will now show on a separate line
the Bottom Material and thickness.