Create fractional strings for thickness, different from dimstr().
New function thickness_str(x) is like dimstr(x), except that its uses a
granularity of 1/4s instead of 1/16s, and it does not produce `strong' or `shy'
indications. It just rounds in the usual way to the nearest fourth.

Change all places in the Job class where thickness measurements are output to use
thickness_str() instead of dimstr().