Fix material string size / spacing in panel drawings.
Change marine-grade plywood abbreviation from 'MARPLY' to 'MarPLY' to save space.

Calculate width of material abbreviation string in points, and if the amount
that the panel rectangle is wider than the string is less than 6 points, use
half of the extra space as the right-hand padding, rather than 6. This moves the
text right and effectively centers it in the panel rectangle horizontally, which
looks nicer.

Use Helvetica font instead of the default Times-Roman for the material string,
which seems easier to read at the small point sizes.