GUI changes to make leg-mounting section clearer, grayed/disabled when no legs.
The bottom material only appears when legs are being mounted, so it is blank
when there are no legs.

'Bottoms:' label and 'Mounting legs requires bottoms thicker than...' label are
now grayed out until leg-mounting checkbox is checked.

Do not show 0.0 in bottom thickness entry box when btmpanel1 and btmpanel2
thicknesses are empty.