Fix bug that sometimes results in too-narrow or too-wide fillers.
Added module-level constants for min_filler_width and max_filler_width.

Changed how the cabinet_width property is calculated, to keep the filler
widths within the allowable range.

Also changed how the extra_width property is calculated, so that is now
depends on the cabinet_width.

Added setter and deleter methods for the fullwidth property.