Cleaned up much lint in job.py, fixed imports, set __all__, yapf reformatting.
Many docstrings were fixed to conform to PEP257 standards.

Yapf reformatted a bunch of long lines that had operators at the beginning of
next lines, moving them to the end like the old usage. Doesn't one of the PEPs
specify that operators should be moved to after the line break?