Rennamed cabinet-calc subdir to cabinet_calc, added code to modify sys.path.
Cannot name the directory holding the application source files cabinet-calc;
hyphens are not legal in Python identifiers, only underscores.

We added code to the cabcalc executable and the cabinet_calc main module to
add the project root directory to sys.path. This way, absolute imports will
work everywhere in the project.